refactor : 접근제어자
