+++ src/main/java/lotto/Main.java
@@ -1,10 +1,12 @@
 package lotto;
 
+import lotto.domain.Lotto;
+import lotto.utils.Utils;
 import lotto.view.InputView;
 import lotto.view.OutputView;
 
 public class Main {
-    static Lotto lotto = new Lotto();
+    private static Lotto lotto = new Lotto();
 
     public static void main(String[] args) {
         int money = Utils.purchase(InputView.inputMoney());
