+++ src/main/java/lotto/utils/Utils.java
@@ -1,4 +1,4 @@
-package lotto;
+package lotto.utils;
 
 import java.util.ArrayList;
 import java.util.Collections;
